backend/requirements.txt
flask
flask-cors
pinecone
rank_bm25
spacy
groq
python-docx
pypdf
python-pptx
requests
python-dotenv
werkzeug
google-generativeai
